All medium files split such that we do not have a giant file for each medium, but a folder
