+++ include/medium/compute_wavefield.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "acoustic_isotropic2d/wavefield.hpp"
-#include "elastic_anisotropic2d/elastic_anisotropic2d.hpp"
+#include "elastic_anisotropic2d/wavefield.hpp"
 #include "elastic_isotropic2d/wavefield.hpp"
 #include <Kokkos_Core.hpp>
