+++ include/medium/elastic_anisotropic2d/properties.hpp
