+++ include/medium/elastic_anisotropic2d/source.hpp
@@ -1,0 +1,42 @@
+#pragma once
+
+#include "algorithms/dot.hpp"
+#include "algorithms/gradient.hpp"
+#include "enumerations/dimension.hpp"
+#include "enumerations/medium.hpp"
+#include "enumerations/wavefield.hpp"
+#include "globals.h"
+#include "point/field_derivatives.hpp"
+#include "point/properties.hpp"
+#include "point/stress.hpp"
+#include <Kokkos_Core.hpp>
+
+namespace specfem {
+namespace medium {
+
+template <typename PointSourcesType, typename PointPropertiesType>
+KOKKOS_INLINE_FUNCTION auto impl_compute_source_contribution(
+    const std::integral_constant<specfem::dimension::type,
+                                 specfem::dimension::type::dim2>,
+    const std::integral_constant<specfem::element::medium_tag,
+                                 specfem::element::medium_tag::elastic>,
+    const std::integral_constant<specfem::element::property_tag,
+                                 specfem::element::property_tag::anisotropic>,
+    const PointSourcesType &point_source,
+    const PointPropertiesType &point_properties) {
+
+  using PointAccelerationType =
+      specfem::point::field<PointPropertiesType::dimension,
+                            PointPropertiesType::medium_tag, false, false, true,
+                            false, PointPropertiesType::simd::using_simd>;
+
+  PointAccelerationType result;
+
+  result(0) = point_source.stf(0) * point_source.lagrange_interpolant(0);
+  result(1) = point_source.stf(1) * point_source.lagrange_interpolant(1);
+
+  return result;
+}
+
+} // namespace medium
+} // namespace specfem
