+++ include/medium/frechet_derivatives.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "acoustic_isotropic2d/frechet_derivative.hpp"
-#include "elastic_anisotropic2d/elastic_anisotropic2d.hpp"
+#include "elastic_anisotropic2d/frechet_derivative.hpp"
 #include "elastic_isotropic2d/frechet_derivative.hpp"
 #include <Kokkos_Core.hpp>
